hola

+++ src/domain/Ubicacion.java
@@ -2,19 +2,19 @@ package domain;
 
 public class Ubicacion {
 
-	private long latitud;
-	private long longitud;
+	private float latitud;
+	private float longitud;
 	
-	public long getLatitud() {
+	public float getLatitud() {
 		return latitud;
 	}
-	public void setLatitud(long latitud) {
+	public void setLatitud(float latitud) {
 		this.latitud = latitud;
 	}
-	public long getLongitud() {
+	public float getLongitud() {
 		return longitud;
 	}
-	public void setLongitud(long longitud) {
+	public void setLongitud(float longitud) {
 		this.longitud = longitud;
 	}
